# Slight performance boost in `pingora-limit::Estimator`