I was tinkering with the code, to my surprise I made `Estimator` run faster just by using iterators instead of for-loops

The overall diff is about 60% faster on single thread

```Pingora Estimator single thread 2.334566806s total, 23ns avg per operation``` 

vs 

```Pingora Estimator single thread 868.937015ms total, 8ns avg per operation```

There is 2ns gain per thread, from 25ns -> 23ns (8%).


I have attached the complete bench diff for my machine: https://www.diffchecker.com/9EnuDEot/

I may be missing/ignoring something here, cause this is too good to be true.

Though this is a [documented and tested](https://doc.rust-lang.org/book/ch13-04-performance.html) behavior.

I have also added a test for `reset` method, which might be redundant.