@atamakahere-git 2024-03-04
Hey @fee1-dead , I applied your [suggested changes](https://github.com/atamakahere-git/pingora/tree/refactor-suggested-changes), and it seems like `.min()` is loosing performance compared to `.fold()`.
Can you run the benchmark and tell if that's the same case on other machines too?