@fee1-dead 2024-03-04
I don't have much time to run benchmarks myself, but I can trust the results. `fold` doesn't look that bad stylistically when compared to `min`.